requirements.txt
asgiref==3.8.1
beautifulsoup4==4.11.2
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dbfread==2.0.7
dj-database-url==2.0.0
Django==4.2
django-bootstrap4==22.3
django-filter==24.3
django-geojson==4.0.0
django-heroku==0.3.1
django-leaflet==0.28.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
Flask==3.1.0
gunicorn==20.1.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.1.0
psycopg2==2.9.5
requests==2.32.3
soupsieve==2.4
sqlparse==0.4.3
typing_extensions==4.6.3
tzdata==2022.7
urllib3==2.3.0
Werkzeug==3.1.3
whitenoise==6.4.0
wincertstore==0.2
zipp==3.21.0
